Ash's APP1 block inserted

+++ EPII_CM55M_APP_S/app/ww_projects/ww130_cli/fatfs_task.h
@@ -19,34 +19,36 @@
 #include "metadata.h"
 
 // Experimental: set a limit on the name of files
-#define	FNAMELEN 16
+#define FNAMELEN 16
 
 // The states for the fatfs_task
 // APP_FATFS_STATE_NUMSTATES is only used to establish the number of states
-typedef enum {
-	APP_FATFS_STATE_UNINIT						=0x0000,
-	APP_FATFS_STATE_IDLE						=0x0001,
-	APP_FATFS_STATE_BUSY						=0x0002,
-	APP_FATFS_STATE_NUMSTATES					=0x0003
+typedef enum
+{
+	APP_FATFS_STATE_UNINIT = 0x0000,
+	APP_FATFS_STATE_IDLE = 0x0001,
+	APP_FATFS_STATE_BUSY = 0x0002,
+	APP_FATFS_STATE_NUMSTATES = 0x0003
 } APP_FATFS_STATE_E;
 
 // Structure to use for file operations:
 // Initially for reading and writing a file
-typedef struct {
-	char *		fileName;
-	uint8_t *	buffer;		// Pointer to the buffer containing file contents
-	uint32_t 	length;		// Number of bytes to write or read
-	FRESULT 	res;		// Result code returned from fatFs
-	ImageMetadata * metadata;	// Metadata to write to the file
-	bool		closeWhenDone;	// If true the file is closed when the operation completes
-	bool		unmountWhenDone;	// If true the SD card is unmounted when the operation completed
-	QueueHandle_t senderQueue;	// FreeRTOS queue that will get the response
+typedef struct
+{
+	char *fileName;
+	uint32_t *buffer;		   // Pointer to the buffer containing file contents
+	uint32_t length;		   // Number of bytes to write or read
+	FRESULT res;			   // Result code returned from fatFs
+	ImageMetadata *metadata;   // Metadata to write to the file
+	bool closeWhenDone;		   // If true the file is closed when the operation completes
+	bool unmountWhenDone;	   // If true the SD card is unmounted when the operation completed
+	QueueHandle_t senderQueue; // FreeRTOS queue that will get the response
 } fileOperation_t;
 
 TaskHandle_t fatfs_createTask(int8_t priority);
 
 uint16_t fatfs_getState(void);
 
-const char * fatfs_getStateString(void);
+const char *fatfs_getStateString(void);
 
 #endif /* APP_WW_PROJECTS_WW130_CLI_FATFS_TASK_H_ */
